move switch chain logic into `cancelOrder` util
